:white_check_mark: add test case for message.config (#21556)
